refactor: Change page URLs
