+++ client/src/components/AppHeader.vue
@@ -41,7 +41,7 @@
             :aria-label="$t('header.account_dropdown_aria')"
             data-cy="headerUserMenu"
           >
-            <q-item clickable to="/account/metadata">
+            <q-item clickable to="/account/profile">
               <q-item-section avatar>
                 <q-icon name="account_circle" />
               </q-item-section>
@@ -49,7 +49,7 @@
                 {{ $t("header.profile") }}
               </q-item-section>
             </q-item>
-            <q-item clickable data-cy="link_my_account" to="/account/profile">
+            <q-item clickable data-cy="link_my_account" to="/account/login">
               <q-item-section avatar>
                 <q-icon name="o_settings" />
               </q-item-section>
