+++ client/src/router/routes.js
@@ -43,8 +43,8 @@ const routes = [
             component: () => import("src/pages/Account/ProfilePage.vue"),
           },
           {
-            path: "metadata",
-            component: () => import("src/pages/Account/MetadataPage.vue"),
+            path: "login",
+            component: () => import("src/pages/Account/LoginPage.vue"),
           },
         ],
       },
