+++ client/test/cypress/integration/copywriting.cy.js
@@ -26,8 +26,8 @@ describe("Copywriting Screenshots", () => {
         screenshotPath('/register')
         screenshotPath('/dashboard')
         screenshotPath("/publications", "[data-cy=publications_list]")
-        screenshotPath("/account/profile", "[data-cy=update_user_email]")
-        screenshotPath("/account/metadata", "[data-cy=position_title]")
+        screenshotPath("/account/profile", "[data-cy=position_title]")
+        screenshotPath("/account/login", "[data-cy=update_user_email]")
         screenshotPath("/publication/1/setup/basic", "[data-cy=visibility_field]")
         screenshotPath("/publication/1/setup/users", "[data-cy=editors_list]")
         screenshotPath("/publication/1/setup/content", "[data-cy=content_block_select]")
